Allow "Identity=default" in .pkla files
(or, to be more precise,
   Identity=unix-group:foo;default;...
as well.)

It's tempting to suggest
   Identity=unix-group:*
in a low-priority .pkla file as an alternative; unfortunately
pklocalauthority(8) is quite explicit that the files are processed one
group at a time, so
   Identity=unix-group:foo
would override
   Identity=unix-group:*
only if "foo" were the last evaluated group regardless of priorities.

Having an user/group/default structure is anyway moving us a bit closer
to an ordinary ACL model (but not close enough, we still operate within
the existing semantis framework).